change main modals (not using yet)
